news-summarizer/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "news-summariser"
version = "0.1.0"
description = "AI-powered news summariser with sentiment and cost observability"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openai>=1.12.0",
  "cohere==5.20.4",
  "requests>=2.31.0",
  "python-dotenv>=1.0.0",
  "tiktoken>=0.5.0",
  "gradio>=4.44.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "ruff>=0.6.0",
]

[project.scripts]
news-summariser = "news_summariser.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"
